posix: Introduce flags for validity of iatt members
v1 of the patch started off as adding new fields to iatt that can be
filled up using statx but the discussions were more around introducing
masks to check the validity of different fields from a RIO perspective.
To that extent, I have dropped the statx call in this version and
introduced a 64 bit mask for existing fields. The masks I have defined
are similar with the statx() flags' masks.

I have *not* changed iatt_to_stat() to use the macros IATT_TYPE_VALID,
IATT_GFID_VALID etc before blindly copying from struct iatt to struct.

Also fixed warnings in xlators because of atime/mtime/ctime seconds
field change from uint32_t to int64_t.

+++ xlators/cluster/ec/src/ec-combine.c
@@ -98,8 +98,8 @@ ec_combine_write (ec_fop_data_t *fop, ec_cbk_data_t *dst,
         return 1;
 }
 
-void ec_iatt_time_merge(uint32_t * dst_sec, uint32_t * dst_nsec,
-                        uint32_t src_sec, uint32_t src_nsec)
+void ec_iatt_time_merge(int64_t *dst_sec, uint32_t *dst_nsec,
+                        int64_t src_sec, uint32_t src_nsec)
 {
     if ((*dst_sec < src_sec) ||
         ((*dst_sec == src_sec) && (*dst_nsec < src_nsec)))
